requirements.txt
# ---- Web API & Server ----
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.30.0
gunicorn>=23.0.0

# ---- Data validation (Pydantic v2 stack) ----
pydantic>=2.7.0
pydantic-settings>=2.2.1
email-validator>=2.2.0   # แก้ปัญหา ModuleNotFoundError: email_validator

# ---- Templates & Static files (ถ้าเสิร์ฟ HTML/CSS/JS ผ่าน FastAPI) ----
jinja2>=3.1.4
python-multipart>=0.0.9  # รับฟอร์ม/อัปโหลดไฟล์ (แม้ fastapi[standard] จะมีให้แล้ว แต่ระบุไว้ชัดๆ ก็โอเค)

# ---- Firebase (ถ้าใช้ Auth/Firestore ผ่าน backend) ----
firebase-admin>=6.5.0
google-cloud-firestore>=2.16.0

# ---- Password hashing ----
passlib[bcrypt]>=1.7.4   # รวม passlib + bcrypt ในตัว

# ---- Util ----
python-dotenv>=1.0.1     # โหลดค่า ENV จาก .env
httpx>=0.27.0            # เรียก API ภายนอก (ถ้าจำเป็น)
